move logic out of CountrySelect component into VPN page
